+++ examples/simple/source/posts.generator.rb
@@ -1,0 +1,11 @@
+
+set :template, "post"
+
+data.blog.each do |post|
+  # mdt= Tilt['markdown'].new { post.content }
+  # post.body = mdt.render 
+
+  page "#{post.slug}.html", :template=>'post', :post=>post
+end
+
+#page "index.html"